requirements.txt
langchain
fastapi
langgraph
langchain_community
langchain_core
langchain_groq
uvicorn
watchdog
python-dotenv
langgraph-cli[inmem]